expenses/permissions.py
from rest_framework import permissions
from authentication.models import User
class IsOwner(permissions.BasePermission):

    def has_object_permission(self, request, view, obj):
        return obj.owner == request.user