requirements.txt
asgiref==3.8.1
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.5
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Faker==24.4.0
filelock==3.13.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
platformdirs==4.2.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.25.1
